backend/requirements.txt
# Core dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-socketio>=5.10.0
python-multipart>=0.0.6
pydantic>=2.5.0
aiofiles>=23.2.1

# CORS and middleware
python-cors>=1.7.0

# Environment variables
python-dotenv>=1.0.0

# HTTP requests
httpx>=0.25.2
requests>=2.31.0

# JSON processing
orjson>=3.9.10

# Logging
loguru>=0.7.2

# Integration with vibe-worldbuilding-mcp
# This will be installed via: pip install -e ../../vibe-worldbuilding-mcp
# mcp>=0.1.0

# Development dependencies (install separately)
# black>=23.11.0
# pylint>=3.0.3
# mypy>=1.7.1
# pytest>=7.4.3
# pytest-asyncio>=0.21.1 